backend/requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
joblib==1.0.1
krwordrank==1.0.3
numpy==1.20.3
psutil==5.8.0
pytz==2021.1
scikit-learn==0.24.2
scipy==1.6.3
soynlp==0.0.493
sqlparse==0.4.1
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
